requirements.txt
certifi==2020.6.20
colorama==0.4.4
importlib-resources==5.4.0
pdf2image==1.16.0
Pillow==8.4.0
PyPDF3==1.0.6
pytesseract==0.3.8
tqdm==4.64.0
wincertstore==0.2
zipp==3.6.0
